Try factoring even with missing rs
With a bit of luck (like in the new test case) we can find a product of
rs that is square even when we have fewer than L+1 rs.

We do this by adding a bunch of 1s to our list of rs and keeping the
requirement that the resulting product can't be 1.